+++ MovieRecommendationAPI/Controllers/Movie/MovieController.cs
@@ -1,5 +1,8 @@
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using MovieRecommendation.Dtos.Movie;
+using MovieRecommendation.Services.Auth;
 using MovieRecommendation.Services.Movie;
 
 namespace MovieRecommendation.Controllers.Movie;
@@ -9,10 +12,27 @@ namespace MovieRecommendation.Controllers.Movie;
 public class MovieController : ControllerBase
 {
     private readonly IMovieService _movieService;
+    private readonly IUserService _userService;
 
-    public MovieController(IMovieService movieService)
+    public MovieController(IMovieService movieService, IUserService userService)
     {
         _movieService = movieService;
+        _userService = userService;
+    }
+
+    private Guid? GetCurrentUserId()
+    {
+        var userIdString = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (userIdString == null) return null;
+        return Guid.Parse(userIdString);
+    }
+
+    private async Task<bool> IsAdmin()
+    {
+        var userId = GetCurrentUserId();
+        if (userId == null) return false;
+        var user = await _userService.GetUserByIdAsync((Guid)userId);
+        return user != null && user.IsAdmin;
     }
 
     [HttpGet]
@@ -26,28 +46,34 @@ public class MovieController : ControllerBase
     public async Task<IActionResult> GetMovieById(Guid id)
     {
         var movieDto = await _movieService.GetMovieByIdAsync(id);
-        if(movieDto == null)  return NotFound();
+        if (movieDto == null) return NotFound();
         return Ok(movieDto);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<IActionResult> CreateMovie(CreateMovieDto createMovieDto)
     {
+        if (!(await IsAdmin())) return Unauthorized();
         var movieDto = await _movieService.CreateMovieAsync(createMovieDto);
         return Ok(movieDto);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<IActionResult> UpdateMovie(Guid id, UpdateMovieDto updateMovieDto)
     {
-        var movieDto = await _movieService.UpdateMovieAsync(id,  updateMovieDto);
-        if(movieDto == null) return NotFound();
+        if (!(await IsAdmin())) return Unauthorized();
+        var movieDto = await _movieService.UpdateMovieAsync(id, updateMovieDto);
+        if (movieDto == null) return NotFound();
         return Ok(movieDto);
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<IActionResult> DeleteMovie(Guid id)
     {
+        if (!(await IsAdmin())) return Unauthorized();
         await _movieService.DeleteMovieByIdAsync(id);
         return NoContent();
     }
